+++ src/actors/actor.h
@@ -6,8 +6,8 @@
 // 位置やサイズ、色などの基本的な情報を持つ
 class Actor {
 private:
-  u_int        uid_;       // 固有識別id
-  u_int        uniqueId(); // 固有idを生成する
+  uintmax_t    uid_;       // 固有識別id
+  uintmax_t    uniqueId(); // 固有idを生成する
   
   bool         finishSetup_;
   bool         enableUpdate_;
